clk: Ingenic: Fix bugs related to SSI clock.
The SSI clock of X1000 are not directly derived from the
output of SSIPLL, but from the clock obtained by dividing
the frequency by 2. "X1000_CLK_SSIPLL_DIV2" is added for
this purpose. And to ensure that the clocks are initialized
correctly, the ABIs in "x1000-cgu.h" are reordered.

Signed-off-by: 周琰杰 (Zhou Yanjie) <zhouyanjie@wanyeetech.com>